refactor: web reads the daemon's Issue types instead of copying them
`Issue`, `IssueStatus` and `RepoIssues` were declared three times: the
daemon's issues-store plus a hand-copy in each of lib/types.ts and
lib/issues.ts. The daemon publishes the shape structurally unchanged over
`issue.snapshot`, so a field it adds silently never reaches the web types.

Both web modules now re-export the store's declaration, and board.ts takes
its Issue from lib/issues.ts alongside every other issue consumer. The
"a live task here hides the issue from the unified board" rule moves onto
`isLinkedIssue`, which is where it is implemented. Type-only.
